move some exceptions into common
